switch away from --env=disable
So, for some reason, when --env=disable, there are `process.env` checks in the bundled code that runs in the dashboard. I really didn't want to have to set NODE_ENV=production by default because of the interaction with BETTER_AUTH_SECRET being required, but it's better than things being broken.